# labels/cidr: Memoize labels for already seen prefixes