--- a/pkg/labels/cidr/cidr_test.go
+++ b/pkg/labels/cidr/cidr_test.go
@@ -4,8 +4,11 @@
 package cidr
 
 import (
+	"math/rand"
 	"net/netip"
 	"runtime"
+	"strconv"
+	"sync"
 	"testing"
 
 	. "github.com/cilium/checkmate"
@@ -314,6 +317,43 @@ func BenchmarkLabels_SortedListCIDRIDs(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		_ = lbls.SortedList()
+	}
+}
+
+func BenchmarkGetCIDRLabelsConcurrent(b *testing.B) {
+	prefixes := make([]netip.Prefix, 0, 16)
+	octets := [4]byte{0, 0, 1, 1}
+	for i := 0; i < 16; i++ {
+		octets[0], octets[1] = byte(rand.Intn(256)), byte(rand.Intn(256))
+		prefixes = append(prefixes, netip.PrefixFrom(netip.AddrFrom4(octets), 32))
+	}
+
+	for _, goroutines := range []int{1, 2, 4, 16, 32, 48} {
+		b.Run(strconv.Itoa(goroutines), func(b *testing.B) {
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				b.StopTimer()
+				start := make(chan struct{})
+				var wg sync.WaitGroup
+
+				wg.Add(goroutines)
+				for j := 0; j < goroutines; j++ {
+					go func() {
+						defer wg.Done()
+
+						<-start
+
+						for k := 0; k < 64; k++ {
+							_ = GetCIDRLabels(prefixes[rand.Intn(len(prefixes))])
+						}
+					}()
+				}
+
+				b.StartTimer()
+				close(start)
+				wg.Wait()
+			}
+		})
 	}
 }