GetCIDRLabels turns a CIDR into a set of labels representing the cidr itself and all broader CIDRS which include the specified CIDR in them.

In case of ToFQDNs policies matching many names that maps to many IPs with very short TTLs, this function can easily become a hot path.

To get better performance, it is possible to memoize the results from previous computations. This is useful since the function computes the labels for all broader CIDRs which include the one passed as argument, so the memoization should be especially effective for all these cases.

```
name                             old time/op    new time/op    delta
GetCIDRLabels/0.0.0.0/0-8           271ns ±26%     329ns ±36%     ~     (p=0.190 n=10+10)
GetCIDRLabels/10.16.0.0/16-8       4.36µs ±11%    2.16µs ±11%  -50.38%  (p=0.000 n=10+10)
GetCIDRLabels/192.0.2.3/32-8       8.54µs ±10%    4.13µs ±11%  -51.67%  (p=0.000 n=10+9)
GetCIDRLabels/192.0.2.3/24-8       6.91µs ±11%    3.24µs ±10%  -53.06%  (p=0.000 n=10+10)
GetCIDRLabels/192.0.2.0/24-8       6.93µs ±13%    3.27µs ±12%  -52.76%  (p=0.000 n=10+10)
GetCIDRLabels/::/0-8                325ns ± 8%     305ns ± 0%   -6.14%  (p=0.000 n=8+7)
GetCIDRLabels/fdff::ff/128-8       50.6µs ± 9%    16.3µs ±16%  -67.86%  (p=0.000 n=10+10)
GetCIDRLabels/f00d:42::ff/128-8    53.6µs ± 9%    15.9µs ±12%  -70.24%  (p=0.000 n=10+10)
GetCIDRLabels/f00d:42::ff/96-8     42.1µs ± 7%    12.2µs ± 8%  -70.92%  (p=0.000 n=10+10)

name                             old alloc/op   new alloc/op   delta
GetCIDRLabels/0.0.0.0/0-8            656B ± 0%      720B ± 0%   +9.76%  (p=0.000 n=10+10)
GetCIDRLabels/10.16.0.0/16-8       3.17kB ± 0%    3.29kB ± 0%   +3.78%  (p=0.000 n=8+10)
GetCIDRLabels/192.0.2.3/32-8       6.88kB ± 0%    6.80kB ± 0%   -1.16%  (p=0.000 n=10+7)
GetCIDRLabels/192.0.2.3/24-8       6.32kB ± 0%    6.21kB ± 0%   -1.77%  (p=0.000 n=10+10)
GetCIDRLabels/192.0.2.0/24-8       6.32kB ± 0%    6.21kB ± 0%   -1.77%  (p=0.000 n=10+10)
GetCIDRLabels/::/0-8                 656B ± 0%      720B ± 0%   +9.76%  (p=0.000 n=10+10)
GetCIDRLabels/fdff::ff/128-8       25.9kB ± 0%    25.0kB ± 0%   -3.39%  (p=0.000 n=7+10)
GetCIDRLabels/f00d:42::ff/128-8    28.8kB ± 0%    25.0kB ± 0%  -13.03%  (p=0.000 n=10+10)
GetCIDRLabels/f00d:42::ff/96-8     25.1kB ± 0%    23.4kB ± 0%   -6.86%  (p=0.000 n=10+10)

name                             old allocs/op  new allocs/op  delta
GetCIDRLabels/0.0.0.0/0-8            3.00 ± 0%      3.00 ± 0%     ~     (all equal)
GetCIDRLabels/10.16.0.0/16-8         37.0 ± 0%       3.0 ± 0%  -91.89%  (p=0.000 n=10+10)
GetCIDRLabels/192.0.2.3/32-8         69.0 ± 0%       3.0 ± 0%  -95.65%  (p=0.000 n=10+10)
GetCIDRLabels/192.0.2.3/24-8         53.0 ± 0%       3.0 ± 0%  -94.34%  (p=0.000 n=10+10)
GetCIDRLabels/192.0.2.0/24-8         53.0 ± 0%       3.0 ± 0%  -94.34%  (p=0.000 n=10+10)
GetCIDRLabels/::/0-8                 3.00 ± 0%      3.00 ± 0%     ~     (all equal)
GetCIDRLabels/fdff::ff/128-8          449 ± 0%         4 ± 0%  -99.11%  (p=0.000 n=10+10)
GetCIDRLabels/f00d:42::ff/128-8       449 ± 0%         4 ± 0%  -99.11%  (p=0.000 n=10+10)
GetCIDRLabels/f00d:42::ff/96-8        295 ± 0%         4 ± 0%  -98.64%  (p=0.000 n=10+10)
```