@pippolo84 2023-10-17
Integration Tests hit https://github.com/cilium/cilium/issues/27080