@odinuge 2023-10-24
Hi!

It does seem like this introduces some strange behavior. My colleague @sjdot discovered that this change results in excessive numbers of false positive policy drops, for legit traffic that should not be dropped.


By adding this test

```diff
diff --git a/pkg/labels/cidr/cidr_test.go b/pkg/labels/cidr/cidr_test.go
index 8ee0dbd605..9a6dd722c4 100644
--- a/pkg/labels/cidr/cidr_test.go
+++ b/pkg/labels/cidr/cidr_test.go
@@ -4,6 +4,7 @@
 package cidr

 import (
+       "fmt"
        "net/netip"
        "testing"

@@ -309,6 +310,18 @@ func BenchmarkLabels_SortedListCIDRIDs(b *testing.B) {
        }
 }

+func TestThisThing(t *testing.T) {
+       for _, ip := range []string{
+               "10.0.0.0/1",
+               "10.0.0.0/3",
+       } {
+               t.Run(ip, func(b *testing.T) {
+                       lbls := GetCIDRLabels(netip.MustParsePrefix(ip))
+                       fmt.Printf("%q: %+v\n", netip.MustParsePrefix(ip), lbls.LabelArray())
+               })
+       }
+}
+
 func BenchmarkIPStringToLabel(b *testing.B) {
        for _, ip := range []string{
                "0.0.0.0/0",
  ```
  , we get the following output;
  
  ```
  === RUN   TestThisThing
=== RUN   TestThisThing/10.0.0.0/1
"10.0.0.0/1": [cidr:0.0.0.0/0 cidr:0.0.0.0/1 reserved:world-ipv4]
=== RUN   TestThisThing/10.0.0.0/3
"10.0.0.0/3": [cidr:0.0.0.0/0 cidr:0.0.0.0/1 reserved:world-ipv4]
--- PASS: TestThisThing (0.00s)
    --- PASS: TestThisThing/10.0.0.0/1 (0.00s)
    --- PASS: TestThisThing/10.0.0.0/3 (0.00s)
PASS
```

When disabling the cache, we instead get this that make more sense.

```
=== RUN   TestThisThing
=== RUN   TestThisThing/10.0.0.0/1
"10.0.0.0/1": [cidr:0.0.0.0/0 cidr:0.0.0.0/1 reserved:world-ipv4]
=== RUN   TestThisThing/10.0.0.0/3
"10.0.0.0/3": [cidr:0.0.0.0/0 cidr:0.0.0.0/1 cidr:0.0.0.0/2 cidr:0.0.0.0/3 reserved:world-ipv4]
--- PASS: TestThisThing (0.00s)
    --- PASS: TestThisThing/10.0.0.0/1 (0.00s)
    --- PASS: TestThisThing/10.0.0.0/3 (0.00s)
PASS
```

So it overall seems like there is some issue in the way things are cached here, or is this expected..? @pippolo84 @joamaki