@pippolo84 2023-10-24
Hi @odinuge , thank you for the bringing this up. :pray: 
I confirm this was not expected. We are working on a fix, together with better unit testing.